mejorado visualizacion de la pagina y arreglando pequeños bugs

+++ code/barbershopApp/context_processor.py
@@ -1,6 +1,6 @@
 def total_carrito(request):
-    total = 0
+    total = 0.0
     if "carrito" in request.session.keys():
         for key, value in request.session["carrito"].items():
-            total += int(value["precio"])
+            total += float(value["precio"])
     return {"total_carrito" : total}
